fix alternatives parsing
