+++ libnavigation-base/build.gradle
@@ -29,7 +29,7 @@ dependencies {
     implementation dependenciesList.mapboxNavigator
 
     implementation dependenciesList.androidXAnnotation
-
+    implementation dependenciesList.coroutinesAndroid
     implementation dependenciesList.kotlinStdLib
 
     testImplementation project(':libtesting-utils')
