+++ libnavigation-core/src/main/java/com/mapbox/navigation/core/routealternatives/RouteAlternativesController.kt
@@ -1,13 +1,19 @@
 package com.mapbox.navigation.core.routealternatives
 
 import com.mapbox.api.directions.v5.models.DirectionsRoute
+import com.mapbox.base.common.logger.model.Message
+import com.mapbox.base.common.logger.model.Tag
 import com.mapbox.navigation.base.route.RouteAlternativesOptions
 import com.mapbox.navigation.base.route.RouterOrigin
+import com.mapbox.navigation.base.utils.parseDirectionsResponse
 import com.mapbox.navigation.core.directions.session.DirectionsSession
 import com.mapbox.navigation.core.directions.session.RoutesExtra.ROUTES_UPDATE_REASON_ALTERNATIVE
 import com.mapbox.navigation.core.trip.session.TripSession
 import com.mapbox.navigation.navigator.internal.MapboxNativeNavigator
+import com.mapbox.navigation.utils.internal.logI
+import kotlinx.coroutines.runBlocking
 import java.util.concurrent.CopyOnWriteArraySet
+import java.util.concurrent.TimeUnit
 
 internal class RouteAlternativesController constructor(
     private val options: RouteAlternativesOptions,
@@ -19,7 +25,7 @@ internal class RouteAlternativesController constructor(
         .apply {
             setRouteAlternativesOptions(
                 com.mapbox.navigator.RouteAlternativesOptions(
-                    options.intervalMillis * SECONDS_PER_MILLIS,
+                    TimeUnit.MILLISECONDS.toSeconds(options.intervalMillis).toDouble(),
                     MIN_TIME_BEFORE_MANEUVER_SECONDS,
                     LOOK_AHEAD_SECONDS
                 )
@@ -66,13 +72,19 @@ internal class RouteAlternativesController constructor(
             changedRoutes.add(currentRoute)
 
             // Map the alternatives from nav-native, add the existing RouteOptions.
-            val alternatives = routeAlternatives.map { routeAlternative ->
-                DirectionsRoute.fromJson(
-                    routeAlternative.route,
-                    currentRoute.routeOptions(),
-                    null // We don't know the requestUuid at this point
-                )
+            var alternatives: List<DirectionsRoute>
+            // TODO make async
+            runBlocking {
+                alternatives = routeAlternatives.map { routeAlternative ->
+                    parseDirectionsResponse(
+                        routeAlternative.route,
+                        currentRoute.routeOptions()
+                    ) {
+                        logI(TAG, Message("Response metadata: $it"))
+                    }.first()
+                }
             }
+
             changedRoutes.addAll(alternatives)
 
             directionsSession.setRoutes(
@@ -96,8 +108,8 @@ internal class RouteAlternativesController constructor(
     }
 
     private companion object {
+        private val TAG = Tag("MbxRouteAlternativesController")
         private const val MIN_TIME_BEFORE_MANEUVER_SECONDS = 1.0
         private const val LOOK_AHEAD_SECONDS = 1.0
-        private const val SECONDS_PER_MILLIS = 0.001
     }
 }
